Finalized functions to a functional point

+++ backend/src/v1/Functions/Products/RemoveProduct.js
@@ -7,13 +7,21 @@ module.exports = async (req, res) => {
   
     const target = req.params.id;
   
+    if (!target ) {
+      return res.status(400).send({error: 'Missing product ID'});
+    }
+  
+    if (!auth) {
+      return res.status(401).send({error: 'Unauthorized'});
+    }
+
     try {
       const connection = await database.connect();
   
       const [user] = await connection.query('SELECT * FROM `users` WHERE UserID = ?;', [Number.parseInt(auth)]);
       if (user[0].PriviledgeID !== 0) {
         await connection.end();
-        return res.status(401).send('Unauthorized');
+        return res.status(401).send({error: 'Unauthorized'});
       }
       
       const [ordersConstraint] = await connection.query('DELETE FROM `orders` WHERE ProductID = ?;', [Number.parseInt(target)]);
@@ -23,7 +31,7 @@ module.exports = async (req, res) => {
       await res.send(result[0]);
     } catch (error) {
       console.error('Error deleting user:', error);
-      await res.status(500).send('Error deleting user');
+      await res.status(500).send({error: 'Error deleting user'});
     }
 
 };
